+++ git-mailmap/src/lib.rs
@@ -7,6 +7,8 @@ pub fn parse(buf: &[u8]) -> parse::Lines<'_> {
     parse::Lines::new(buf)
 }
 
+mod entry;
+
 #[derive(PartialEq, Eq, Debug, Hash, Ord, PartialOrd, Clone, Default)]
 #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
 pub struct Entry<'a> {
@@ -18,52 +20,4 @@ pub struct Entry<'a> {
     commit_name: Option<&'a BStr>,
     /// The email to look for and replace.
     commit_email: Option<&'a BStr>,
-}
-
-mod entry {
-    use crate::Entry;
-    use bstr::BStr;
-
-    impl<'a> Entry<'a> {
-        pub fn change_name_by_email(proper_name: impl Into<&'a BStr>, commit_email: impl Into<&'a BStr>) -> Self {
-            Entry {
-                canonical_name: Some(proper_name.into()),
-                commit_email: Some(commit_email.into()),
-                ..Default::default()
-            }
-        }
-        pub fn change_email_by_email(proper_email: impl Into<&'a BStr>, commit_email: impl Into<&'a BStr>) -> Self {
-            Entry {
-                canonical_email: Some(proper_email.into()),
-                commit_email: Some(commit_email.into()),
-                ..Default::default()
-            }
-        }
-        pub fn change_name_and_email_by_email(
-            proper_name: impl Into<&'a BStr>,
-            proper_email: impl Into<&'a BStr>,
-            commit_email: impl Into<&'a BStr>,
-        ) -> Self {
-            Entry {
-                canonical_name: Some(proper_name.into()),
-                canonical_email: Some(proper_email.into()),
-                commit_email: Some(commit_email.into()),
-                ..Default::default()
-            }
-        }
-
-        pub fn change_name_and_email_by_name_and_email(
-            proper_name: impl Into<&'a BStr>,
-            proper_email: impl Into<&'a BStr>,
-            commit_name: impl Into<&'a BStr>,
-            commit_email: impl Into<&'a BStr>,
-        ) -> Self {
-            Entry {
-                canonical_name: Some(proper_name.into()),
-                canonical_email: Some(proper_email.into()),
-                commit_name: Some(commit_name.into()),
-                commit_email: Some(commit_email.into()),
-            }
-        }
-    }
 }
