+++ git-mailmap/src/parse.rs
@@ -37,6 +37,14 @@ impl<'a> Iterator for Lines<'a> {
     type Item = Result<Entry<'a>, Error>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        todo!()
+        for line in self.lines.by_ref() {
+            self.line_no += 1;
+            match line.first() {
+                None => continue,
+                Some(b) if *b == b'#' => continue,
+                Some(_) => {}
+            }
+        }
+        None
     }
 }
